% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalue.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{Calcula el p-valor}
\usage{
calculate_pvalue(statistic = 1, n = 10, prob_density = c("z", "t"),
  tails = c("right", "left", "two"))
}
\arguments{
\item{statistic}{Estadistico de contraste estandarizado}

\item{n}{Tamaño de la muestra}

\item{prob_density}{Distribucion a utilizar \code{c("z", "t")}}

\item{tails}{Tipo de prueba a realizar \code{c("right", "left", "two")}}
}
\value{
Un vector de tipo numerico con el valor p
}
\description{
Con base a los parametros de la funcion se calcula el p-valor
}
\details{
Esta funcion se calcula para poder incorporar el p-value dentro del
  grafico.
}
\examples{
calculate_pvalue(statistic = 3.08, n = 50, prob_density = "z", tails =
"right")

}
