% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring_font.R
\name{coloring_font}
\alias{coloring_font}
\title{Aplica un color de fuente a un texto en rmarkdowm}
\usage{
coloring_font(x, color)
}
\arguments{
\item{x}{Cadena de texto que se requiere colorear}

\item{color}{Un color entre comillas de texto}
}
\value{
La cadena de texto con el color especificado
}
\description{
Se puede apicar tanto a LaTex como a HTML
}
\details{
Lo que hace la función es aplicar la función en Latex y
 la etiqueta de font color en HTML.
}
\examples{
color_fuente(x = "hola mundo", color = "blue")
}
