% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Calcula el intervalo de confianza para un determinado nivel de confianza}
\usage{
confidence_interval(x, cl, n, s, tails = 2)
}
\arguments{
\item{x}{Media muestral}

\item{cl}{Nivel de confianza}

\item{n}{Tamano de la muestra}

\item{s}{Desviación estándar de la poblacion}

\item{tails}{Desviacion estándar de la poblacion}
}
\value{
Un vector de longitud 2 con el intervalo
}
\description{
Funciona solo cuando se conoce la varianza de la poblacion.
}
\details{
Esta funcion solo aplica para calcular el intervalo cuando se conoce
la desviacion estandar de la poblacion, ya que de lo contrario requerira que
se utilice el valor critico utilizando la distribucion t.
}
\examples{
calcular_intervalo(x = 100, cl = 0.95, n = 50, s = 3, tails = 2)

}
