% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_quantile.R
\name{critical_quantile}
\alias{critical_quantile}
\title{Calcula el cuantíl crítico}
\usage{
critical_quantile(cl = 0.95, tails = 2)
}
\arguments{
\item{cl}{Nivel de confianza}

\item{tails}{Cantidad de colas}
}
\value{
Un vector de longitud 1 con el nivel de confianza ajustado
}
\description{
Funciona para una distribución de dos colas
}
\details{
Esta función solo aplica para calcular el cuantíl crítico en el caso
de una prueba de dos colas, ya que en una prueba de una cola solo es necesario
ingresar el nivel de confianza directamente.
}
\examples{
cuantil_critico(cl = 0.90, tails = 2)

}
