% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_matrix.R
\name{density_matrix}
\alias{density_matrix}
\title{Crear una matriz con base a una funcion de densidad especificada}
\usage{
density_matrix(distr = "t", n = 20)
}
\arguments{
\item{distr}{EDistribucion a utilizar \code{c("z", "t")}}

\item{n}{Tamano de la muestra}
}
\value{
Un tibble con dos columnas
}
\description{
Sombrea la region de rechazo para una prueba de una cola
}
\details{
Esta funcion grafica utilizando la distribucion t. El parametro
\code{lower_tail} dependera de si la prueba es positiva derecha o negativa.  En
este caso el planteamiento de la hipotesis alternativa sera lo que defina el
lado de la cola bajo estudio. Recordar que, por convencion, siempre nos
interesara sombrear el área de rechazo y dejar la region de aceptacion sin
sombrear.
}
\examples{
density_matrix(distr = "t", n = 30)

}
