% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{Obtiene la etiqueta para el titulo}
\usage{
get_label(density_function, tipo_prueba)
}
\arguments{
\item{density_function}{Distribucion a utilizar \code{c("z", "t")}}

\item{tipo_prueba}{Tipo de prueba a realizar \code{c("right", "left", "two")}}
}
\value{
Un vector de tipo caracter
}
\description{
En funcion de los parametros \code{density_function} y \code{tipo_prueba}
}
\details{
Esta funcion unicamente obtiene la etiqueta correcta para titulo del
  grafico
}
\examples{
get_label("t", "right")

}
