% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test.R
\name{plot_test}
\alias{plot_test}
\title{Realiza pruebas de hipotesis para distribuciones t y z}
\usage{
plot_test(sample_size = 10, test_statistic = 1, critical_value = 1.5,
  density_function = c("z", "t"), tipo_prueba = c("right", "left",
  "two"))
}
\arguments{
\item{sample_size}{Tama침o de la muestra}

\item{test_statistic}{Estadistico de contraste estandarizado}

\item{critical_value}{Valor critico dado un nivel de confianza}

\item{density_function}{Distribucion a utilizar \code{c("z", "t")}}

\item{tipo_prueba}{Tipo de prueba a realizar \code{c("right", "left", "two")}}
}
\value{
Un gr치fico en ggplot
}
\description{
Sombrea la regi칩n de rechazo para pruebas de una y dos colas
}
\details{
Esta funci칩n grafica en funcion de los parametros de entrada una
  distribucion \strong{z} o una \strong{t} y ejecuta el tipo de prueba
  seleccionado, el cual puede ser \strong{right}, \strong{left} y \strong{two}
}
\section{Aviso}{
 La funcion requiere una validacion de argumentos mas rigurosa.
}

\examples{
plot_test(sample_size = 110, test_statistic = 2.52, critical_value = 1.645,
density_function = "t", tipo_prueba = "two")

}
