% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_to_fill.R
\name{region_to_fill}
\alias{region_to_fill}
\title{Sombrea las regiones de la distribucion}
\usage{
region_to_fill(density_function, tipo_prueba, critical_value, sample_size,
  intervalo)
}
\arguments{
\item{density_function}{Distribucion a utilizar \code{c("z", "t")}}

\item{tipo_prueba}{Tipo de prueba a realizar \code{c("right", "left", "two")}}

\item{critical_value}{Valor critico dado un nivel de confianza}

\item{sample_size}{Tamaño de la muestra}

\item{intervalo}{El rango de los limetes de sombreado a nivel de eje x}
}
\value{
Un objeto de tipo ggplot
}
\description{
Rellan con un color en funcion de la distribucion seleccionada y del tipo de
prueba
}
\details{
Esta función aplica un relleno de color en funcion de si es una
  distribucion t o z, asi como si es una prueba unilateral o bilateral.
}
\examples{
plot_test(sample_size = 110, test_statistic = 2.52, critical_value = 1.645,
density_function = "t", tipo_prueba = "two")

}
