% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_yunkel.R
\name{theme_yunkel}
\alias{theme_yunkel}
\title{Tema especifico para graficar distribuciones}
\usage{
theme_yunkel()
}
\value{
Un gráfico en ggplot
}
\description{
Se construye a partir de cowplot
}
\details{
Define espacios superior para una mejor impresion sobre rmarkdown
}
\examples{
plot_test(sample_size = 110, test_statistic = 2.52, critical_value = 1.645,
density_function = "t", tipo_prueba = "two")

}
