% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_onetail.R
\name{ttest_onetail}
\alias{ttest_onetail}
\title{Grafica una distribución t para una cola}
\usage{
ttest_onetail(n, t, t_alfa, lower_tail = FALSE)
}
\arguments{
\item{t}{Estadístico t (t-score)}

\item{t_alfa}{Valor crítico}
}
\value{
Un gráfico en ggplot
}
\description{
Sombrea la región de rechazo para una prueba de una cola
}
\details{
Esta función grafica utilizando la distribución t. El parámetro
`lower_tail` dependerá de si la prueba es positiva (derecha) o negativa.  En
este caso el planteamiento de la hipótesis alternativa será lo que defina el
lado de la cola bajo estudio. Recordar que, por convención, siempre nos
interesará sombrear el área de rechazo y dejar la región de aceptación sin
sombrear.
}
\examples{


}
