% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_twotails.R
\name{ttest_twotails}
\alias{ttest_twotails}
\title{Grafica una distribución t para dos colas}
\usage{
ttest_twotails(n, t_score, nivel_confianza = 0.95)
}
\arguments{
\item{t}{Estadístico t (t-score)}

\item{t_alfa}{Valor crítico}
}
\value{
Un gráfico en ggplot
}
\description{
Sombrea la región de rechazo para una prueba de una cola
}
\details{
Esta función grafica utilizando la distribución t
}
\examples{

}
