% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest_twotails.R
\name{ztest_twotails}
\alias{ztest_twotails}
\title{Grafica una distribución normal estandarizada}
\usage{
ztest_twotails(z_alfa, z)
}
\arguments{
\item{z_alfa}{Valor crítico en escala estandarizada}

\item{z}{Estadístico de prueba}
}
\value{
Un gráfico en ggplot
}
\description{
Sombrea la región de rechazo para una prueba de dos colas
}
\details{
Esta función grafica utilizando la distribución normal estandarizada,
considerando como valores de entrada un valor crítico pre-calculado y el estadístico de
prueba z.
}
\examples{
zdist_plot_twotail(z_alfa = 1.96, z = 1)

}
