% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fuente.R
\name{color_fuente}
\alias{color_fuente}
\title{Aplica un color de fuente a un texto en rmarkdowm}
\usage{
color_fuente(x, color)
}
\arguments{
\item{x}{Cadena de texto que se requiere colorear}

\item{color}{Un color entre comillas de texto}
}
\value{
La cadena de texto con el color especificado
}
\description{
Se puede apicar tanto a LaTex como a HTML
}
\details{
Lo que hace la función es aplicar la función `\\textcolor` en Latex y
 la etiqueta de font color en HTML.
}
\section{convencion}{
 esta es una extensión de la explicación que se debe de
 dar sobre esta funcion.  \eqn{a + b}, así como \deqn{a + b}

 \deqn{H_{1}: \mu_{0} > 7}
}

\examples{
color_fuente("hola mundo", "blue")

}
